refactor: 당첨 로또 개수  Long으로 범위 확장

+++ src/main/java/lotto/controller/LottoController.java
@@ -84,7 +84,7 @@ public class LottoController {
         }
     }
 
-    private void displayResult(Map<Prize, Integer> prizeCounts, double earnRate) {
+    private void displayResult(Map<Prize, Long> prizeCounts, double earnRate) {
         outView.printLottoResult(prizeCounts);
         outView.printEarnRate(earnRate);
     }
